+++ VersionHistory.md
@@ -1,6 +1,6 @@
 # Version History
 
-## 1.2 -- TBD
+## 1.2.0 -- 2 Apr 2026
 
 | Issue No. | Change Type | Description |
 |--------|--------|-------|
@@ -10,6 +10,9 @@
 | NA     | Fix    | The application was not properly setting the payload type in RTP packets if the application sent a re-INVITE to add media to a call. |
 | NA     | Change | The Answer button in the main form now displays the Call Form if a ringing or Auto-Answered call is picked up and put on-line. |
 | NA     | Fix    | The user was allowed to put a ringing call on-hold. |
+| NA     | Change | Updated the AudoAnswer.jpg static image file. |
+| NA     | Change | Updated to use verion 1.0.1 of the SipLib.Video.Window DLL.|
+| NA     | Change | Limit the video resolution to VGA for capture so that the application can run reliably on older versions of Windows with low performance hardware on low bandwidth networks when media encryption is enabled. |
 
 ## v1.1.0 - 20 Feb 2026
 | Issue No. | Change Type | Description |